+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/api/service/ServiceCallBackResourceImpl.kt
@@ -1,11 +1,10 @@
-package com.tencent.devops.process.api
+package com.tencent.devops.process.api.service
 
 import com.tencent.devops.common.api.pojo.Page
 import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.api.util.PageUtil
 import com.tencent.devops.common.pipeline.event.CallBackEvent
 import com.tencent.devops.common.web.RestResource
-import com.tencent.devops.process.api.user.UserCallBackResource
 import com.tencent.devops.process.engine.service.ProjectPipelineCallBackService
 import com.tencent.devops.process.pojo.ProjectPipelineCallBack
 import com.tencent.devops.process.pojo.ProjectPipelineCallBackHistory
@@ -13,9 +12,9 @@ import com.tencent.devops.process.pojo.pipeline.enums.CallBackNetWorkRegionType
 import org.springframework.beans.factory.annotation.Autowired
 
 @RestResource
-class UserCallBackResourceImpl @Autowired constructor(
+class ServiceCallBackResourceImpl @Autowired constructor(
     val projectPipelineCallBackService: ProjectPipelineCallBackService
-) : UserCallBackResource {
+) : ServiceCallBackResource {
 
     override fun create(
         userId: String,
