+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/service/ProjectPipelineCallBackUrlGenerator.kt
@@ -1,6 +1,8 @@
 package com.tencent.devops.process.engine.service
 
+import com.tencent.devops.process.pojo.pipeline.enums.CallBackNetWorkRegionType
+
 interface ProjectPipelineCallBackUrlGenerator {
 
-    fun generateCallBackUrl(url: String): String
+    fun generateCallBackUrl(region: CallBackNetWorkRegionType?, url: String): String
 }
