+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/service/ProjectPipelineCallBackUrlGeneratorImpl.kt
@@ -1,8 +1,12 @@
 package com.tencent.devops.process.engine.service
 
+import com.tencent.devops.process.pojo.pipeline.enums.CallBackNetWorkRegionType
+import org.springframework.stereotype.Service
+
+@Service
 class ProjectPipelineCallBackUrlGeneratorImpl : ProjectPipelineCallBackUrlGenerator {
 
-    override fun generateCallBackUrl(url: String): String {
+    override fun generateCallBackUrl(region: CallBackNetWorkRegionType?, url: String): String {
         return url
     }
 }
